Added navigation bar, bootstrap styles for forms, and other small upgrades.

+++ app.py
@@ -2,6 +2,7 @@ from flask import Flask # Para usar flask
 from flask import render_template # Para redireccionar una u otra URL
 from flask import request # Para tomar datos de formularios etc.
 from flask import redirect # Para redirigir a la misma pagina (o a otra? no se)
+from flask import send_from_directory
 from flaskext.mysql import MySQL # Para que Flask pueda conectarse con sql
 from datetime import datetime # Para tomar la fecha y hora de la computadora, la usamos para cambiar el nombre de archivo de la imagen
 import os
@@ -18,6 +19,9 @@ mysql.init_app(app)
 CARPETA = os.path.join('uploads') # Se ingresa la ruta del directorio que quiero referenciar
 app.config['CARPETA'] = CARPETA # Cuando en el sistema hago referencia a la palabra CARPETA, hago referencia al directorio uploads
 
+@app.route('/uploads/<nombreFoto>') # Cuando alguien quiere levantar una foto, el codigo lo va a reconocer aca y va a ejecutar el codigo
+def uploads(nombreFoto):
+    return send_from_directory(app.config['CARPETA'], nombreFoto)
 
 @app.route('/') # Lo que sucede al entrar al index
 def index():  # ese 'index' puede tener otro nombre que yo quiera.
@@ -26,7 +30,6 @@ def index():  # ese 'index' puede tener otro nombre que yo quiera.
     cursor = conn.cursor() # Se crea un cursor
     cursor.execute(sql) # Hago que el cursor ejecute la consulta
     empleados = cursor.fetchall() # Trae todos los datos y los mete en una tupla
-    print(empleados)
     conn.commit() # Se envian los cambios a la base de datos.
     return render_template('empleados/index.html', empleados=empleados) # En el segundo parametro le pongo el nombre de la variable que quiero
 
@@ -39,7 +42,9 @@ def destroy(id):
     #fila = cursor.fetchall() # Se obtienen los datos
     #os.remove(os.path.join(app.config['CARPETA'], fila[0][0])) # Se elimina la imagen guardada en la carpeta uploads
     fila = cursor.fetchone() # Se obtienen los datos. fetchone obtiene un solo dato, por lo que es bueno usarlo cuando el id es unico. Ver que en otra uso fetchall, eso trae mas datos y hay que poner fila[0][0]
-    os.remove(os.path.join(app.config['CARPETA'], fila[0]))
+
+    if fila[0] != "":
+        os.remove(os.path.join(app.config['CARPETA'], fila[0]))
 
     cursor.execute("DELETE FROM empleados WHERE id=%s", id)
     conn.commit()
@@ -51,7 +56,6 @@ def edit(id):
     cursor = conn.cursor()
     cursor.execute("SELECT * FROM empleados WHERE id=%s", id) # Solo se busca el empleado con el id que se ingreso
     empleados = cursor.fetchall()
-    print(empleados)
     conn.commit()
     return render_template('empleados/edit.html', empleados=empleados) # Envia al usuario a /edit
 
@@ -61,12 +65,14 @@ def update():
     _nombre = request.form['txtNombre']
     _correo = request.form['txtCorreo']
     _foto = request.files['txtFoto']
+    print(_foto.filename)
     sql = 'UPDATE empleados SET nombre=%s, correo=%s WHERE id=%s;'
     datos = (_nombre, _correo, id)
     conn = mysql.connect()
     cursor = conn.cursor()
     now = datetime.now()
     tiempo = now.strftime("%Y%H%M%S")
+    nuevoNombreFoto = ''
 
     if _foto.filename != '':
         nuevoNombreFoto = tiempo + _foto.filename
@@ -94,20 +100,24 @@ def storage():
 
     now = datetime.now()
     tiempo = now.strftime("%Y%H%M%S")
+
+    nuevoNombreFoto = ''
     if _foto.filename != '': # Se verifica que la foto se haya cargado. Si no, nombre es vacio.
         nuevoNombreFoto = tiempo + _foto.filename
         _foto.save("uploads/" + nuevoNombreFoto)
+    else:
+        nuevoNombreFoto = ""
 
     sql =  "INSERT INTO `empleados` (`id`, `nombre`, `correo`, `foto`) VALUES (NULL, %s, %s, %s);"
     
     datos = (_nombre, _correo, nuevoNombreFoto)
     
-    # Grabar en la BD (Base de Datos) VVVV
+    # Carga en la BD (Base de Datos) VVVV
     conn = mysql.connect()
     cursor = conn.cursor()
     cursor.execute(sql, datos) # Se ejecuta la consulta en mysql, y el segundo parametro es la lista de datos con los que se reemplazan %s del string
     conn.commit()
-    return render_template('empleados/index.html') # Al terminar, vuelve a mandar al usuario al index
+    return redirect('/') # Redirige a la app a @app.route('/') y hace todo lo que está ahi (busca en la BD y carga eso en el index)
 
 if __name__ == '__main__':
     app.run(debug=True) # Ejecuta la aplicacion y comienza a correr el controlador
